+++ sync_flutter_libs/lib/objectbox_sync_flutter_libs.dart
@@ -1,2 +1,9 @@
-/// This package only contains platform-specific native libraries for flutter.
-/// All the dart code is in package "objectbox".
+/// This package contains platform-specific native libraries for flutter.
+/// See the actual library implementation in package "objectbox".
+
+import 'dart:io';
+
+import 'package:path_provider/path_provider.dart';
+
+Future<Directory> defaultStoreDirectory() async =>
+    await getApplicationDocumentsDirectory();